+++ src/utils/cn.js
@@ -1,4 +1,4 @@
-nom import { clsx } from 'clsx';
+import { clsx } from 'clsx';
 import { twMerge } from 'tailwind-merge';
 
 /**